# src/pages/api/v1/openapi/logins.json/logins.json
{
    "openapi": "3.0.2",
    "paths": {
        "/logins": {
            "get": {
                "tags": [
                    "login"
                ],
                "summary": "Returns all items",
                "description": "Returns a list of items",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200ListOfItems"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                }
            },
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Add a new item",
                "description": "Add a new item",
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginCreateBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200SingleItem"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                }
            }
        },
        "/logins/{id}": {
            "get": {
                "tags": [
                    "login"
                ],
                "summary": "Find item by ID",
                "description": "Returns a single item",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idParam"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200SingleItem"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            },
            "patch": {
                "tags": [
                    "login"
                ],
                "summary": "Update an item",
                "description": "",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idParam"
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/LoginUpdateBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200SingleItem"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    }
                }
            },
            "delete": {
                "tags": [
                    "login"
                ],
                "summary": "Delete an item",
                "description": "",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idParam"
                    }
                ],
                "responses": {
                    "204": {
                        "$ref": "#/components/responses/204"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        }
    },
    "components": {
        "hiddenSchemas": {
            "LoginBaseModel": {
                "type": "object",
                "required": true,
                "properties": {
                    "version": {
                        "type": "integer",
                        "format": "int32",
                        "example": 1
                    },
                    "lastModifiedDate": {
                        "type": "string",
                        "example": "1970-01-01T00:00:00.0Z"
                    }
                }
            },
            "LoginEditableModel": {
                "type": "object",
                "required": true,
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "foo"
                    },
                    "path": {
                        "type": "string",
                        "example": "/"
                    },
                    "username": {
                        "type": "string",
                        "example": "demo"
                    },
                    "secret": {
                        "type": "string",
                        "example": "one234"
                    },
                    "otp": {
                        "type": "string",
                        "example": "oneTimePassword"
                    },
                    "note": {
                        "type": "string",
                        "example": "This is my item"
                    }
                }
            }
        },
        "schemas": {
            "LoginModel": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/hiddenSchemas/LoginBaseModel"
                    },
                    {
                        "$ref": "#/components/hiddenSchemas/LoginEditableModel"
                    }
                ]
            }
        },
        "parameters": {
            "idParam": {
                "name": "id",
                "in": "path",
                "description": "ID of item",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        },
        "requestBodies": {
            "LoginCreateBody": {
                "description": "Create request body",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "logins"
                                        },
                                        "attributes": {
                                            "$ref": "#/components/hiddenSchemas/LoginEditableModel"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "LoginUpdateBody": {
                "description": "Update request body",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "bSzkmX7Zec7sOV2atzn3r"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "logins"
                                        },
                                        "attributes": {
                                            "$ref": "#/components/hiddenSchemas/LoginEditableModel"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "responses": {
            "200SingleItem": {
                "description": "Successful operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "bSzkmX7Zec7sOV2atzn3r"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "logins"
                                        },
                                        "attributes": {
                                            "$ref": "#/components/schemas/LoginModel"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "200ListOfItems": {
                "description": "Successful operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "bSzkmX7Zec7sOV2atzn3r"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "logins"
                                            },
                                            "attributes": {
                                                "$ref": "#/components/schemas/LoginModel"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "204": {
                "description": "Successful operation"
            },
            "401": {
                "description": "Authentication failed"
            },
            "404": {
                "description": "Item not found"
            },
            "422": {
                "description": "Unprocessable Entity"
            }
        }
    }
}